Update and rename script assignemnt 3.py to Commodity price discovery using natural variables: An application of statistical and Machine Learning models to Cocoa’s future prices
This project consists of an empirical analysis on cocoa future prices consideringas drivers non-financial variables: cocoa is a commodity that is widely traded in financial markets, and its price can be influenced by various factors. Among those, changes in weather patterns can affect cocoa production and lead to supply disruptions, that can in turn greatly influence the price of cocoa.
Historical data concerning natural variables are collected and chosen to be specific to where production is located and with respect to cocoa farming idiosyncrasies.
Those are analysed singularly together with the historical future
prices of cocoa traded in the US.
Afterwards, quantitative models are implemented to evaluate their relationships to check if the selected variables can be used to predict the movement offuture prices. 
This is useful to provide an indication for hedging climate risks in cocoa markets and to understand the dynamics of commodity prices with respect to weather-related risk.